pyproject.toml
[project]
name = "cdm-spark-events"
version = "0.1.0"
description = "Event processor for the KBase CDM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "delta-spark==3.3.1",
    "kafka-python==2.2.4",
    "pandas==2.2.3",
    "pydantic-settings==2.9.1",
    "pyspark==3.5.5",
    "python-json-logger==3.3.0",
    "pyyaml>=6.0.2",
    "requests==2.32.3",
]

[dependency-groups]
dev = [
    "ipython>=9.2.0",
    "pytest==8.3.5",
    "pytest-cov==6.1.1",
]
